fix(decision-log): use userBizKey instead of internal userID, fix e2e test scripts
- Changed CreatedBy from internal uint ID to int64 bizKey so creatorName
  resolution works via FindByBizKeys lookup
- Fixed e2e test role creation to use valid permissionCodes ('team:read')
- Added beforeAll diagnostics to catch undefined IDs early
- Fixed task file just test-e2e syntax (positional arg, no --flag)
- Added lesson doc on e2e beforeAll failure cascade misdiagnosis

+++ backend/internal/handler/router_test_stubs.go
@@ -358,16 +358,16 @@ var _ service.ProgressService = (*StubProgressSvc)(nil)
 // StubDecisionLogSvc is a stub DecisionLogService for test wiring.
 type StubDecisionLogSvc struct{}
 
-func (s *StubDecisionLogSvc) Create(_ context.Context, _ int64, _ uint, _ dto.DecisionLogCreateReq) (*model.DecisionLog, error) {
+func (s *StubDecisionLogSvc) Create(_ context.Context, _ int64, _ int64, _ dto.DecisionLogCreateReq) (*model.DecisionLog, error) {
 	return nil, fmt.Errorf("stub: not implemented")
 }
-func (s *StubDecisionLogSvc) Update(_ context.Context, _ int64, _ uint, _ dto.DecisionLogUpdateReq) (*model.DecisionLog, error) {
+func (s *StubDecisionLogSvc) Update(_ context.Context, _ int64, _ int64, _ dto.DecisionLogUpdateReq) (*model.DecisionLog, error) {
 	return nil, fmt.Errorf("stub: not implemented")
 }
-func (s *StubDecisionLogSvc) Publish(_ context.Context, _ int64, _ uint) (*model.DecisionLog, error) {
+func (s *StubDecisionLogSvc) Publish(_ context.Context, _ int64, _ int64) (*model.DecisionLog, error) {
 	return nil, fmt.Errorf("stub: not implemented")
 }
-func (s *StubDecisionLogSvc) List(_ context.Context, _ int64, _ uint, _ dto.Pagination) (*dto.PageResult[model.DecisionLog], error) {
+func (s *StubDecisionLogSvc) List(_ context.Context, _ int64, _ int64, _ dto.Pagination) (*dto.PageResult[model.DecisionLog], error) {
 	return nil, fmt.Errorf("stub: not implemented")
 }
